minor bug fix.

+++ DataHandling/DataHandling.c
@@ -59,7 +59,7 @@ int get_format_version(int *version, char *blue_spike_data_path)
 		if (!(get_word_in_line('\t', 1, word, line, TRUE)))
 			return 0;
 		fclose(fp);   	
-		*version = atof(word);
+		*version = (int)atof(word);
 		return 1;
 	}
 	return 0;
